Stop tracking .clangd
It's up to the user to define how clangd should work on their machine.
That's because paths are annoying to work with in clangd.
Example of .clangd where the LSP is ask to look at the current file as
if it were compiling it with -Wall -Wextra and with a library located at
  /absolute/path/to/lib/headers:

```
CompileFlags:
    Add: [-Wall, -Wextra, -std=c99, -I/absolute/path/to/lib/headers]
```