Changing how shaders work to be more flexible.
Previously, I had to update the shader context struct if I wanted to
create a new behavior in my shader...
This is not flexible at all.

Solution: Now a shader is a struct which has 2 elements:
  - A function `run`: `Color (*run)(int x, int y, void* context) which
    is the actual code of the shader. This function is meant to be called on
    each pixel related to the parent function.
    - A void ptr to a context. This allows to send a context with variable
    data size to the shader `run` function. In the `run` function
    implementations, the first thing that has to be done is to cast the
    `void* context` to a pointer of the proper size.

Note: Due to C not having `static_cast` capabilities, the compiler
cannot check at compile time that the casting is allowed.
So we can totally send whatever pointer to a `run` shader function and
we will only see the errors at runtime.